+++ src/actions/settingsActions.ts
@@ -4,27 +4,27 @@
  * SPDX-License-Identifier: LicenseRef-Nordic-4-Clause
  */
 
-import Store from 'electron-store';
-
 import {
     settingsLoad,
     toggleAutoRead as toggleAutoReadAction,
     toggleAutoReset as toggleAutoResetAction,
 } from '../reducers/settingsReducer';
 import { RootState, TDispatch } from '../reducers/types';
-
-const persistentStore = new Store({ name: 'pc-nrfconnect-programmer' });
+import { getSettings, setSettings } from '../store';
 
 export function loadSettings() {
     return (dispatch: TDispatch) => {
-        const settings = persistentStore.get('settings') || {};
-        if (settings.autoRead == null) {
+        const settings = getSettings();
+
+        if (!settings.autoRead) {
             settings.autoRead = false;
         }
+
         if (settings.autoReset == null) {
             settings.autoReset = true;
         }
-        persistentStore.set('settings', settings);
+
+        setSettings(settings);
         dispatch(settingsLoad(settings));
     };
 }
@@ -32,11 +32,11 @@ export function loadSettings() {
 export function toggleAutoRead() {
     return (dispatch: TDispatch, getState: () => RootState) => {
         dispatch(toggleAutoReadAction());
-        const settings = persistentStore.get('settings') || {};
+        const settings = getSettings();
 
         // Do not use async functions above，
         // otherwise the state would be the same as before toggling
-        persistentStore.set('settings', {
+        setSettings({
             ...settings,
             autoRead: getState().app.settings.autoRead,
         });
@@ -46,11 +46,11 @@ export function toggleAutoRead() {
 export function toggleAutoReset() {
     return (dispatch: TDispatch, getState: () => RootState) => {
         dispatch(toggleAutoResetAction());
-        const settings = persistentStore.get('settings') || {};
+        const settings = getSettings();
 
         // Do not use async functions above，
         // otherwise the state would be the same as before toggling
-        persistentStore.set('settings', {
+        setSettings({
             ...settings,
             autoReset: getState().app.settings.autoReset,
         });