+++ src/util/devices.ts
@@ -8,13 +8,12 @@ import nrfdl, {
     DeviceCore,
     DeviceCoreInfo,
     ProtectionStatus,
-    // eslint-disable-next-line import/no-unresolved
 } from '@nordicsemiconductor/nrf-device-lib-js';
 
 import range from './range';
 
 export type CoreDefinition = {
-    name: DeviceCore | string;
+    name: DeviceCore;
     coreNumber: number;
     romBaseAddr: number;
     romSize: number;
@@ -66,7 +65,7 @@ export enum DeviceFamily {
 }
 
 export interface DeviceDefinition {
-    family?: DeviceFamily;
+    family: DeviceFamily;
     type?: string;
     cores: CoreDefinition[];
 }
@@ -74,7 +73,7 @@ export interface DeviceDefinition {
 /**
  * Default definition for device info
  */
-export const deviceDefinition: DeviceDefinition = {
+export const defaultDeviceDefinition: DeviceDefinition = {
     family: DeviceFamily.UNKNOWN,
     type: 'Unknown',
     cores: [defaultCore],
@@ -84,16 +83,29 @@ export const deviceDefinition: DeviceDefinition = {
  * Some information cannot be fetch by @nordicsemiconductor/nrf-device-lib-js
  * Define the default values here
  */
-const deviceDefinitions: DeviceDefinition[] = [
+export const deviceDefinitions: DeviceDefinition[] = [
+    {
+        ...defaultDeviceDefinition,
+        family: DeviceFamily.NRF51,
+        type: 'nRF51822',
+    },
     // nRF52840 dongle cannot fetch info by @nordicsemiconductor/nrf-device-lib-js
     {
-        ...deviceDefinition,
+        ...defaultDeviceDefinition,
         family: DeviceFamily.NRF52,
         type: 'nRF52840',
+        cores: [
+            {
+                ...defaultCore,
+                romSize: 0x100000, // 1 Mb
+                ramSize: 0x40000, // 256 Kb
+                pageSize: 0x1000, // 4Kb
+            },
+        ],
     },
     // nRF9160 has different ficr and uicr base address
     {
-        ...deviceDefinition,
+        ...defaultDeviceDefinition,
         family: DeviceFamily.NRF91,
         type: 'nRF9160',
         cores: [
@@ -106,7 +118,7 @@ const deviceDefinitions: DeviceDefinition[] = [
     },
     // nRF5340 has dual core definitions
     {
-        ...deviceDefinition,
+        ...defaultDeviceDefinition,
         family: DeviceFamily.NRF53,
         type: 'nRF5340',
         cores: [
@@ -131,6 +143,13 @@ const deviceDefinitions: DeviceDefinition[] = [
         ],
     },
 ];
+
+export const coreFriendlyName = (coreName: DeviceCore) =>
+    ({
+        NRFDL_DEVICE_CORE_APPLICATION: 'Application',
+        NRFDL_DEVICE_CORE_MODEM: 'Modem',
+        NRFDL_DEVICE_CORE_NETWORK: 'Network',
+    }[coreName] ?? coreName);
 
 /**
  * Nordic SoftDevice Id referring to pc-nrfutil
@@ -210,81 +229,95 @@ export const JLinkProductIds = [
     ...range(0x1001, 0x101f),
 ];
 
-export const getDeviceDefinition = (type: string): DeviceDefinition =>
-    deviceDefinitions.find((device: DeviceDefinition) =>
+export const getDeviceDefinition = (type: string): DeviceDefinition => {
+    const predefined = deviceDefinitions.find((device: DeviceDefinition) =>
         device?.type?.toLowerCase().includes(type.toLowerCase())
-    ) || {
-        ...deviceDefinition,
-        type,
-    };
-
-// Get device info by calling version command
-export const getDeviceInfoByUSB = (
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    hwInfo: any
-) => {
-    const deviceInfoByUsb = getDeviceDefinition(
-        // hwInfo.deviceVersion example 'NRF52840_AAD0'
-        hwInfo.deviceVersion.slice(0, 8)
     );
-    const [core] = deviceInfoByUsb.cores;
     return {
-        ...deviceInfoByUsb,
-        cores: [
-            {
-                ...core,
-                romSize: hwInfo.romSize,
-                pageSize: hwInfo.romPageSize,
-                ramSize: hwInfo.ramSize,
-            },
-        ],
+        ...(predefined ?? defaultDeviceDefinition),
+        type,
     };
 };
 
+const getProductId = (device: nrfdl.Device) =>
+    parseInt(
+        device.serialPorts.reduce(
+            (m, p) => (p.vendorId === '1915' ? p.productId : '') || m,
+            ''
+        ),
+        16
+    );
+
+const identifyUsbByVersion = (device: nrfdl.Device) => {
+    return device.hwInfo?.deviceVersion?.length > 0
+        ? getDeviceDefinition(
+              device.hwInfo.deviceVersion.slice(0, 8) // example 'NRF52840_AAD0'
+          )
+        : null;
+};
+
+const identifyUsbBySerialPort = (device: nrfdl.Device) => {
+    const productId = getProductId(device);
+
+    // nRF52
+    if (USBProductIds.some(id => id === productId)) {
+        return deviceDefinitions.find(d => d.family === DeviceFamily.NRF52);
+    }
+
+    // nRF91
+    if (ModemProductIds.some(id => id === productId)) {
+        return deviceDefinitions.find(d => d.family === DeviceFamily.NRF91);
+    }
+
+    return null;
+};
+
+// Get device info by calling version command
+export const getDeviceInfoByUSB = (device: nrfdl.Device) => {
+    return (
+        identifyUsbByVersion(device) ||
+        identifyUsbBySerialPort(device) ||
+        defaultDeviceDefinition
+    );
+};
+
 // Get device info by calling @nordicsemiconductor/nrf-device-lib-js
-export const getDeviceInfoByJlink = (device: nrfdl.Device) => {
-    const model = device.jlink.deviceVersion;
-    const family = device.jlink.deviceFamily;
+export const getDeviceInfoByJlink = (
+    device: nrfdl.Device
+): DeviceDefinition => {
+    const type = device.jlink?.deviceVersion ?? 'Unknown';
+    const family = device.jlink?.deviceFamily as DeviceFamily;
 
     return {
-        ...getDeviceDefinition(model),
+        ...getDeviceDefinition(family),
+        type,
         family,
-        cores: [],
-    } as DeviceDefinition;
+    };
 };
 
 /**
  * Add core info to device info
  *
- * @param {DeviceDefinition} deviceInfo - existing device info
+ * @param {DeviceDefinition} existingDefinition - existing device info
+ * @param {number} coreNumber - the number of the core
  * @param {DeviceCoreInfo} inputCoreInfo - core info to be added
- * @param {string} coreName - the name of the core
  * @param {ProtectionStatus} protectionStatus - the protection status
  *
  * @returns {DeviceDefinition} the updated device info
  */
-export const addCoreToDeviceInfo = (
-    deviceInfo: DeviceDefinition,
+export const updateCoreInfo = (
+    existingDefinition: CoreDefinition,
+    coreNumber: number,
     inputCoreInfo: DeviceCoreInfo,
-    coreName: string,
     protectionStatus: ProtectionStatus
-) =>
-    ({
-        ...deviceInfo,
-        cores: [
-            ...deviceInfo.cores,
-            {
-                ...defaultCore,
-                name: coreName,
-                coreNumber: deviceInfo.cores.length,
-                romBaseAddr: inputCoreInfo.codeAddress,
-                romSize: inputCoreInfo.codeSize,
-                pageSize: inputCoreInfo.codePageSize,
-                // TODO: Check if uicrAddress is present in nrfjprog under nrf-device-lib
-                // uicrBaseAddr: inputCoreInfo.uicrAddress,
-                uicrSize: inputCoreInfo.codePageSize,
-                ...inputCoreInfo,
-                protectionStatus,
-            },
-        ],
-    } as DeviceDefinition);
+): CoreDefinition => ({
+    ...existingDefinition,
+    coreNumber,
+    romBaseAddr: inputCoreInfo.codeAddress,
+    romSize: inputCoreInfo.codeSize,
+    pageSize: inputCoreInfo.codePageSize,
+    uicrBaseAddr: inputCoreInfo.uicrAddress,
+    uicrSize: inputCoreInfo.codePageSize,
+    ...inputCoreInfo,
+    protectionStatus,
+});